pyproject.toml
[tool.poetry]
name = "mastodon-welcome"
version = "0.1.0"
description = ""
authors = ["Matthew Scheffel <matt@dataheck.com>"]
readme = "README.md"
packages = [{include = "mastodon_welcome"}]

[tool.poetry.dependencies]
python = "^3.10"
mastodon-py = "^1.5.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
